+++ src/modules/usermodes/showwhois.c
@@ -35,7 +35,7 @@ ModuleHeader MOD_HEADER
 long UMODE_SHOWWHOIS = 0L;
 
 /* Forward declarations */
-int showwhois_whois(Client *requester, Client *target);
+int showwhois_whois(Client *requester, Client *target, NameValuePrioList **list);
 
 MOD_TEST()
 {
@@ -62,7 +62,7 @@ MOD_UNLOAD()
 	return MOD_SUCCESS;
 }
 
-int showwhois_whois(Client *requester, Client *target)
+int showwhois_whois(Client *requester, Client *target, NameValuePrioList **list)
 {
 	if (IsWhois(target) && (requester != target))
 	{